# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Инспектор ИИ — API",
    "version": "1.0.0",
    "description": "Сверка проектной, рабочей и исполнительной документации по матрице из 132 параметров. Асинхронная pull-модель: загрузка возвращает process_id, результат забирается по нему."
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "inspector_session"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "summary": "Лимиты загрузки и хранения",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Изменить лимиты (ТЗ 9.1: 50 МБ на файл, 200 МБ на пакет)",
        "tags": [
          "system"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "max_upload_mb": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 50
                  },
                  "max_package_mb": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 200
                  },
                  "max_pages": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "retention_days": {
                    "type": "integer",
                    "minimum": 1
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "summary": "Уведомления пользователя по роли",
        "tags": [
          "system"
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "unread",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "summary": "Отметить уведомление прочитанным",
        "tags": [
          "system"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "notification_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "summary": "Объекты с цветовой индикацией (модуль 7)",
        "tags": [
          "dashboard"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "section",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "green",
                "yellow",
                "red"
              ]
            },
            "in": "query",
            "name": "color",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "date_from",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "date_to",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}